requirements.txt
aioredis==2.0.1
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==3.2.2
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
cryptography==44.0.2
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
identify==2.6.9
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.1.0
PyYAML==6.0.2
ruff==0.11.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.1
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
virtualenv==20.29.3
